requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-JWT-Extended
bcrypt
flask-cors
Flask-Migrate